services updates after testing

+++ aspnet-core/src/TACHYON.Application.Shared/AddressBook/Dtos/CreateOrEditFacilityDto.cs
@@ -18,12 +18,9 @@ namespace TACHYON.AddressBook.Dtos
         public string Adress { get; set; }
 
 
-        public decimal Longitude { get; set; }
+        public double Longitude { get; set; }
 
-
-        public decimal Latitude { get; set; }
-
-
+        public double Latitude { get; set; }
 
         public int CityId { get; set; }
